criating shutDown
