+++ src/main.ts
@@ -3,13 +3,10 @@ import comment from './helpers/comment'
 import fail from './helpers/fail'
 import { execCommand } from './helpers/execCommand'
 import { formatImage } from './helpers/formatImage'
-import {
-	vercelDeploy,
-	vercelRemoveProjectDeploy,
-	removeSchema,
-} from './tenants/vercel'
+import { vercelDeploy, removeSchema } from './tenants/vercel'
 import prepare from './pipeline/prepare'
 import build from './pipeline/build'
+import shutDown from './pipeline/shutDown'
 
 async function main() {
 	const {
@@ -26,30 +23,17 @@ async function main() {
 	} = await prepare()
 	let { vercelToken, deploymentUrlVercel } = configVercel
 
-	if (teardown && payloadContext.action === 'closed') {
-		try {
-			core.info(`Teardown: ${mountedUrl}`)
-			core.setSecret(surgeToken)
-			await execCommand({
-				command: ['surge', 'teardown', mountedUrl, `--token`, surgeToken],
-			})
-
-			const image = formatImage({
-				buildingLogUrl,
-				imageUrl:
-					'https://user-images.githubusercontent.com/507615/98094112-d838f700-1ec3-11eb-8530-381c2276b80e.png',
-			})
-
-			if (vercelToken) await vercelRemoveProjectDeploy(deploymentUrlVercel)
-
-			return await comment(
-				`:recycle: [PR Preview](https://${outputUrl}) ${gitCommitSha} has been successfully destroyed since this PR has been closed. \n ${image}`
-			)
-		} catch (err) {
-			core.info('teardown error')
-			return await fail(err)
-		}
-	}
+	await shutDown({
+		teardown,
+		payloadContext,
+		mountedUrl,
+		surgeToken,
+		buildingLogUrl,
+		vercelToken,
+		deploymentUrlVercel,
+		outputUrl,
+		gitCommitSha,
+	})
 
 	const deployingImage = formatImage({
 		buildingLogUrl,
